+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## 4.12.2
-*
+
+* [#749](https://github.com/codeborne/selenide/pull/749) Added because method to CollectionCondition -- thanks to Mikhail Sidelnikov @sidelnikovmike
+* [#695](https://github.com/codeborne/selenide/issues/695) Do not open a browser if `Configuration.reopenBrowserOnFail` is `false` and user has not set webdriver manually 
+* Upgrade selenium to 3.13.0 version
+* Upgrade webdrivermanager to 2.2.3 version
 
 ## 4.12.1 (released 02.06.2018)